added transaction page

+++ Backend/user-service/src/main/java/com/finance/security/dto/UserResponse.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class UserResponse {
-    private int userId; // or appropriate type
+    private Long userId; // or appropriate type
     private String username;  // Assuming 'users' is your user class
     private String token;
 }
