+++ Backend/user-service/src/main/java/com/finance/security/repo/CredRepo.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 
 import com.finance.security.model.User;
 @Repository
-public interface CredRepo extends JpaRepository<User, Integer>{
+public interface CredRepo extends JpaRepository<User, Long>{
 
 	public User findByUsername(String username);
